Refactor vaccination schedule responses and details
- Introduced `VaccinationScheduleDetailsResponse` class to encapsulate vaccination schedule details and rounds.
- Removed properties related to vaccination details and rounds from `VaccinationScheduleResponse`.
- Updated `IVaccinationScheduleService` to return `VaccinationScheduleDetailsResponse` in `GetVaccinationSchedule`.
- Modified `GetVaccinationSchedule` in `VaccinationScheduleService` to construct and return the new details response.
- Replaced `GetResponse` with `GetScheduleResponse` to focus on constructing the `VaccinationScheduleResponse`.
